работают социальные сети и хобби

+++ back-end/peoples_world/src/main/java/nsu/ypprpo/peoples_world/repository/SocialNetworksRepository.java
@@ -4,5 +4,5 @@ import nsu.ypprpo.peoples_world.models.SocialNetwork;
 import nsu.ypprpo.peoples_world.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SocialNetworksRepository extends JpaRepository<SocialNetwork, Long> {
+public interface SocialNetworksRepository extends JpaRepository<SocialNetwork, Integer> {
 }
